deepfreeze/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "es-deepfreeze"
version = "1.0.0"
description = "Standalone Elasticsearch S3 Glacier archival and lifecycle management tool"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.8"
authors = [
    { name = "Bret Wortman", email = "bret@wortman.us" },
]
keywords = [
    "elasticsearch",
    "s3",
    "glacier",
    "archival",
    "snapshot",
    "deepfreeze",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Archiving",
    "Topic :: Database",
]
dependencies = [
    "elasticsearch8>=8.0.0",
    "boto3>=1.26.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "voluptuous>=0.14.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "moto[s3]>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
deepfreeze = "deepfreeze.cli.main:cli"

[project.urls]
Homepage = "https://github.com/wortmanb/curator"
Documentation = "https://github.com/wortmanb/curator/tree/standalone/deepfreeze"
Repository = "https://github.com/wortmanb/curator"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.hatch.build.targets.sdist]
include = [
    "deepfreeze/**/*.py",
    "README.md",
    "pyproject.toml",
    "config.yml.example",
]

[tool.black]
line-length = 88
target-version = ["py38", "py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 88
target-version = "py38"
select = ["E", "F", "W", "I", "UP", "B", "C4"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
